server/requirements.txt
flask == 3.0.0
flask_cors == 4.0.0
stepic == 0.5.0
pillow == 10.1.0
gunicorn == 21.2.0